+++ spec/web_image_spec.rb
@@ -13,34 +13,72 @@ describe LinkThumbnailer::WebImage do
 
   subject { foo }
 
+  it { should respond_to :to_a }
   it { should respond_to :to_json }
   it { should respond_to :source_url }
   it { should respond_to :doc }
 
-  describe ".to_json" do
+  describe ".to_a" do
 
     context "with default attributes" do
 
-      let(:attributes) { %w(source_url) }
+      let(:attributes) { [:source_url] }
 
-      subject { foo.to_json }
+      subject { foo.to_a }
 
-      it { JSON.parse(subject).first.keys.should eq(attributes) }
+      it { subject.keys.should eq(attributes) }
 
     end
 
     context "with all attributes" do
 
-      let(:attributes) { %w(source_url mime_type density colums rows filesize number_colors) }
+      let(:attributes) { [:source_url, :mime_type, :density, :colums, :rows, :filesize, :number_colors] }
 
       before do
         attributes.each {|a| foo.class.send(:define_method, a.to_sym) { 'foo' } }
       end
 
+      after do
+        attributes.each {|a| foo.class.send(:undef_method, a.to_sym) }
+      end
+
+      subject { foo.to_a }
+
+      it { subject.keys.should eq(attributes) }
+      it { subject.values.should include('foo') }
+
+    end
+
+  end
+
+  describe ".to_json" do
+
+    context "with default attributes" do
+
+      let(:attributes) { [:source_url] }
+
       subject { foo.to_json }
 
-      it { puts JSON.parse(subject).map {|e| e.keys.first }.should eq(attributes) }
-      it { puts JSON.parse(subject).map {|e| e.values.first }.should include('foo') }
+      it { JSON.parse(subject).keys.map(&:to_sym).should eq(attributes) }
+
+    end
+
+    context "with all attributes" do
+
+      let(:attributes) { [:source_url, :mime_type, :density, :colums, :rows, :filesize, :number_colors] }
+
+      before do
+        attributes.each {|a| foo.class.send(:define_method, a.to_sym) { 'foo' } }
+      end
+
+      after do
+        attributes.each {|a| foo.class.send(:undef_method, a.to_sym) }
+      end
+
+      subject { foo.to_json }
+
+      it { JSON.parse(subject).keys.map(&:to_sym).should eq(attributes) }
+      it { JSON.parse(subject).values.should include('foo') }
 
     end